Integrate mainline sDDF and makefile snippet + example
This sDDF example was using a copies of the sDDF components from the
serial subsystem. This isn't great, especially with how the sDDF might
change, so it was instead updated to use an sDDF submodule. Since the
sDDF is not structured in a way that is suitable for CMake, the build
for this example was rewritten in GNU Make and a top-level Makefile
snippet which can be included in other projects was added.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>